backend/backend/requirements.txt
flask
diffusers
torch
transformers
accelerate
safetensors
gtts
Pillow
python-dotenv
opencv-python
